Fixed the game sessions test and added code to catch errors in async funcs
Initially the async function "game.run()" was returning but game.gameFinished
was false, when we were asserting it to be true.
As Duncan was saying, this is how async functions are supposed to work - unless
you "await" for them, they return immediately when called, and the async work
happens later in the event loop.

When I added back 'async function()' and 'await game.run()' into the test code,
the behaviour was it would hang, and the test would timeout with no done() being
called.
It turns out this was because there was an error in game.run() (the base NACView
class did not have a log() method). This threw an exception, BUT EXCEPTIONS ARE
SILENTLY SWALLOWED INSIDE ASYNC FUNCTIONS!
I have added try/catch blocks so the errors are now visible.